เพิ่มทศนิยม ทั้งหมด

+++ database/migrations/2020_10_16_085806_create_location_products_table.php
@@ -18,7 +18,7 @@ class CreateLocationProductsTable extends Migration
             $table->integer('location_model_id');
             $table->string('name',50)->unique();
             $table->text('img_path')->nullable();
-            $table->integer('price')->default(0);
+            $table->float('price')->default(0);
             $table->string('description')->nullable();
             $table->timestamps();
         });
